chore(eslint): change package, move files, update imports
