+++ src/routes/dashboard/dashboard.tsx
@@ -15,7 +15,8 @@ import AccountOverview from '@/routes/dashboard/account-overview/account-overvie
 import { Overview } from '@/routes/dashboard/account-overview/account-overview.types';
 import AddTransaction from '@/routes/dashboard/add-transaction/add-transaction';
 import Budget from '@/routes/dashboard/budget/budget';
-import { Category, Table } from '@/routes/dashboard/budget/budget.types';
+import { Table } from '@/routes/dashboard/budget/budget.types';
+import { Categories } from '@/routes/dashboard/categories/categories.types';
 import ManageCategories from '@/routes/dashboard/manage-categories/manage-categories';
 import SavingGoals from '@/routes/dashboard/saving-goals/saving-goals';
 import { SavingGoalsFormProps } from '@/routes/dashboard/saving-goals/saving-goals.types';
@@ -29,7 +30,7 @@ interface DashboardProps {
   budget: Table[];
   saving_goals: SavingGoalsFormProps[];
   overview: Overview[];
-  categories: Category[];
+  categories: Categories;
   upcoming_payment: UpcomingPayment[];
 }
