+++ src/routes/dashboard/saving-goals/saving-goals.tsx
@@ -9,14 +9,13 @@ import { useDashboard } from '@/hooks/use-dashboard';
 import { useModal } from '@/hooks/use-modal';
 
 import { ButtonPrimary, ButtonSecondary } from '@/components/button';
-import { Input } from '@/components/forms/custom-form';
+import { Input } from '@/components/custom-form';
 import Icon from '@/components/icon';
 import Modal from '@/components/modal';
 import SavingGoalsItem from '@/routes/dashboard/saving-goals/saving-goals-item';
 import { SavingGoalsFormProps, SavingGoalsProps } from '@/routes/dashboard/saving-goals/saving-goals.types';
 
 import { ERROR_MSG } from '@/data/errorMessages';
-
 
 export default function SavingGoals(data: SavingGoalsProps) {
   const [savingGoalList, setSavingGoalList] = useState<SavingGoalsProps>(data);
